add bottomsheet for generic info
